fix bug 1. use u8 in TCB syscall_times: [u8; MAX_SYSCALL_NUM]  ,instead u32

+++ os3/src/task/task.rs
@@ -10,7 +10,7 @@ pub struct TaskControlBlock {
     pub task_status: TaskStatus,
     pub task_cx: TaskContext,
     // LAB1: Add whatever you need about the Task.
-    pub syscall_times: [u32;MAX_SYSCALL_NUM],
+    pub syscall_times: [u8;MAX_SYSCALL_NUM],
     pub start_time: usize,
     pub started: bool,
 }
